pyproject.toml
[project]
name = "fofa-agent"
version = "0.1.0"
description = "Threat intel agent for ClickFix discovery via Fofa"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openrouter>=0.2.1",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
    "requests>=2.32.0",
]
